requirements.txt
django==3.2.18
django-appconf==1.0.5
django-ckeditor==6.5.1
django-filter==22.1
django-imagekit==4.1.0
django-qrcode==0.3
django-tastypie==0.14.5
fcm-django==1.0.12
mysqlclient
olefile==0.45.1
pilkit==2.0
Pillow==9.4.0
pytz==2018.3
six==1.11.0
